fix(server): container/network
- container: Create/Inspect 适配新协议
- container: Create 生成UUID, 记录数据库
- container: Create 内部UUID写入labels 和 env
- container: Create 创建容器检查镜像是否存在, 磁盘使用限制写入labels
- container: List返回磁盘使用信息
- container: InspectTemplate 模板详情接口
- image: 镜像列表返回镜像大小
- log interceptor: remove controller debug log of agent reply.
- 添加容器网络访问权限控制
- 修改创建网络脚本为交互式
- 添加节点网络访问权限控制
- 修改镜像管理镜像仓库推送检查
- 添加节点和容器网络访问权限控制
- 修改容器网络配置

Fix #53553

+++ backend/server/agent/internal/cadvisor_helper.go
@@ -47,7 +47,7 @@ func getContainerStats() (map[string]*pb.ResourceStat, error) {
 
 			p = &pb.ResourceStat{
 				MemStat: &pb.MemoryStat{
-					Used: tail.Memory.WorkingSet / megaBytes,
+					Used: float64(tail.Memory.WorkingSet) / megaBytes,
 				},
 			}
